+++ aura/CHANGELOG.md
@@ -25,7 +25,7 @@ is all while modernizing the code and seeing a ~17% decrease in overall code siz
   Great for debugging.
 - **Flag:** `-Br` has been restored as short-hand for `-B --restore`.
 - **Flag:** `-Cb` added as a short-hand for `-C --backup`.
-- **Flag:** The Pacman flags `--cachedir`, `--config`, and `--logfile` also now affect Aura.
+- **Flag:** The Pacman flags `--ignoregroup`, `--cachedir`, `--config`, and `--logfile` also now affect Aura.
 - **Flag:** `--dryrun` no longer requires sudo.
 - Improved Japanese translations thanks to **Onoue Takuro**.
 - Improved Portuguese translations thanks to **Wagner Amaral**.